requirements.txt
asgiref==3.8.1
Django==4.2.13
django-admin-interface==0.28.8
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1